// types/Tprofile.tsx
export interface Tprofile{
    id: number;
    name: string;
    title: string;
    bio: string;
    email: string;
    phone: string;
    address: string;
    profilePic: string;
    linkedIn: string;
    github: string;
    twitter: string;
    facebook: string;
    instagram: string;
    website: string;
    whatsapp: string;
    telegram: string;
    createdAt: string; // ISO date string
    updatedAt: string; // ISO date string
}